readability-implicit-bool-conversion: the rest of the automatic changes

+++ src/cli/main.cpp
@@ -600,7 +600,7 @@ extern "C"
   int rb_hasFile(const char* t_filename) {
     // TODO Consider expanding this to use the path which we have artificially defined in embedded_help.rb
     std::string expandedName = std::string(":/ruby/2.7.0/") + std::string(t_filename) + ".rb";
-    return embedded_files::hasFile(expandedName);
+    return static_cast<int>(embedded_files::hasFile(expandedName));
   }
 
   int rb_require_embedded(const char* t_filename) {
